Fix wanring about signed/unsigned

+++ exif.c
@@ -552,14 +552,14 @@ static void ProcessExifDir(unsigned char * DirStart, unsigned char * OffsetBase,
         if (ByteCount > 4){
             OffsetVal = Get32u(DirEntry+8);
             // If its bigger than 4 bytes, the dir entry contains an offset.
-            if (OffsetVal+ByteCount > ExifLength || OffsetVal > 65536){
+            if (OffsetVal+ByteCount > (unsigned)ExifLength || OffsetVal > 65536){
                 // Bogus pointer offset and / or bytecount value
                 ErrNonfatal("Illegal value pointer for tag %04x in Exif", Tag,0);
                 continue;
             }
             ValuePtr = OffsetBase+OffsetVal;
 
-            if (OffsetVal > ImageInfo.LargestExifOffset){
+            if (OffsetVal > (unsigned)ImageInfo.LargestExifOffset){
                 ImageInfo.LargestExifOffset = OffsetVal;
             }
